Place customized doctests in a separate module.
